Allow defining and calling methods without parentheses
Why do this?

The idea is to allow decoupling how a member is implemented from how it is exposed,
allowing authors to reevaluate the costs of (re-)computation vs. storage
without breaking every existing caller of such code.

This approach is based on experience with earlier attempts, like

- Java's "use getters/setters everywhere, just in case"
- C#'s "use properties, which are source incompatible with getters/setters and binary incompatible with fields"

and provides the desired decoupling while not breaking user code and avoiding other drawbacks of older designs as well.

As the distinction between fun, let and var is nevertheless important,
the idea is that an IDE should use different colors when referring to such members
(which provides more useful information than in Java's or C#'s case).

When to use this?

The idea is that methods which are "constant" for the instance they are called on
are defined without (), and methods that mutate their instance are defined with ().

Example:
  true.not, 1.toString, array.length, iterator.hasNext
        vs.
  iterator.next(), random.nextInt(), ...